+++ src/fhsm_crypto.c
@@ -290,9 +290,10 @@ fhsm_rv_t fhsm_rng_reseed(fhsm_slice_t personalization) {
  * ----------------------------------------------------------------------- */
 size_t fhsm_hash_size(fhsm_hash_t h) {
     switch (h) {
-        case FHSM_HASH_SHA256: case FHSM_HASH_SHA3_256: return 32;
+        case FHSM_HASH_SHA256: case FHSM_HASH_SHA3_256: case FHSM_HASH_SHA512_256: return 32;
         case FHSM_HASH_SHA384: case FHSM_HASH_SHA3_384: return 48;
         case FHSM_HASH_SHA512: case FHSM_HASH_SHA3_512: return 64;
+        case FHSM_HASH_SHA224: case FHSM_HASH_SHA512_224: return 28;
         case FHSM_HASH_SHA1: return 20;
         case FHSM_HASH_MD5:  return 16;
         default: return 0;
@@ -307,6 +308,9 @@ static const char *hash_name(fhsm_hash_t h) {
         case FHSM_HASH_SHA3_256: return "SHA3-256";
         case FHSM_HASH_SHA3_384: return "SHA3-384";
         case FHSM_HASH_SHA3_512: return "SHA3-512";
+        case FHSM_HASH_SHA224:     return "SHA2-224";
+        case FHSM_HASH_SHA512_224: return "SHA512-224";
+        case FHSM_HASH_SHA512_256: return "SHA512-256";
         case FHSM_HASH_SHA1:     return "SHA1";
         case FHSM_HASH_MD5:      return "MD5";
         default:                 return NULL;